// apps/web/src/api/model/listConversations200ItemsItem.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Bens Seguros API
 * OpenAPI spec version: 1.0.0
 */
import type { ListConversations200ItemsItemChannel } from './listConversations200ItemsItemChannel';
import type { ListConversations200ItemsItemContact } from './listConversations200ItemsItemContact';
import type { ListConversations200ItemsItemHandler } from './listConversations200ItemsItemHandler';
import type { ListConversations200ItemsItemStatus } from './listConversations200ItemsItemStatus';

export type ListConversations200ItemsItem = {
  /** @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$ */
  id: string;
  status: ListConversations200ItemsItemStatus;
  handler: ListConversations200ItemsItemHandler;
  assigneeId: string | null;
  contact: ListConversations200ItemsItemContact;
  channel: ListConversations200ItemsItemChannel;
  /**
     * @minimum -9007199254740991
     * @maximum 9007199254740991
     */
  lastSeq: number;
  lastMessageAt: string | null;
  closedAt: string | null;
  /** @pattern ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z))$ */
  createdAt: string;
};